airtable-supabase-sync/requirements.txt
pyairtable>=1.4.0
supabase>=1.0.0
python-dotenv>=0.20.0
requests>=2.28.0
pandas>=1.4.0 